+++ src/php/_helpers/_bc.js
@@ -1,8 +1,8 @@
-module.exports = function _locutus_shared_bc () { // eslint-disable-line camelcase
-  //  discuss at: http://locutusjs.io/php/_locutus_shared/_locutus_shared_bc
+module.exports = function _bc () { // eslint-disable-line camelcase
+  //  discuss at: http://locutusjs.io/php/_helpers/_bc
   // original by: lmeyrick (https://sourceforge.net/projects/bcmath-js/)
   // improved by: Brett Zamir (http://brett-zamir.me)
-  //   example 1: var $bc = _locutus_shared_bc()
+  //   example 1: var $bc = _bc()
   //   example 1: var $result = $bc.PLUS
   //   returns 1: '+'