+++ src/php/_locutus_shared/index.js
@@ -1,1 +1,0 @@
-module.exports['_locutus_shared_bc'] = require('./_locutus_shared_bc')