+++ src/php/bc/bcround.js
@@ -4,7 +4,7 @@ module.exports = function bcround (val, precision) {
   //   example 1: bcround(1, 2)
   //   returns 1: '1.00'
 
-  var _bc = require('../_locutus_shared/_locutus_shared_bc')
+  var _bc = require('../_helpers/_bc')
   var libbcmath = _bc()
 
   var temp, result, digit